Add back colours and text
The .name and .value properties are now on the `d.data` object.

There is also a .value property on the `d` object, but it's
different—the percent calculations are borked with it.
One remaining place does (`d.v || d.value`) but that's because that `d`
already refers to the `.data` property.