Make tooltip contents selectable
This PR makes a few changes to tooltips to make it possible to select
and copy their contents.

The canvas change made it impossible to select the text on the nodes. It
was suggested that an HTML or SVG element could be placed over hovered
nodes to mimick the old selection behaviour, which would be possible but
it could conflict with the panning that also uses mouse dragging.

Instead I changed tooltips so they don't hide when hovered, allowing
users to select the text in there. Tooltips also always contain the full
file path, while the text on nodes is cut off if the nodes are too small.
Additionally the tooltips now position based only on the node position,
not on the cursor, so you can always move the mouse from a node to the
tooltip without hovering over any unrelated pixels that may cause the
tooltip to hide. Also, a delay of 250ms now applies when hiding
tooltips, to make it even easier (have to move at most 18 pixels
so 250ms should be more than enough)