Add node zoom methods
Do `graph.zoom(dataNode)` to zoom in on a node programmatically.

Do `graph.on('zoom', (dataNode) => {})` to listen for node zoom changes.

This allows 0x to implement history management, so it updates the URL hash when a user zooms in on an entry and so 0x can tell the graph to zoom in on a different node when the user goes back in the browser history.