Allow zooming in on a node mid-zoom
Previously it was not possible to pick a different node while a zoom
animation was ongoing. It was not a problem because users could just
wait the few 100ms to click the next node. However, with the
programmatically controlled `graph.zoom()` function, it's feasible to
flick through nodes very quickly; eg when connected to the browser back
button. Without this patch you could end up with an outdated state
because zoom calls were ignored for 500ms.

With this patch, a zoom call mid-zoom cancels the animation and starts
a new animation. It starts it from the current, mid-animation position
as well, so it's nicely smooth! (Except for other stacks popping out of
existence—not quite sure why that happens!)