Fix tooltip position when zoomed using +/- buttons
Previously zooming the entire graph would cause tooltips to show up in
the wrong place. Now, the zoomTransform is applied to the tooltip
position as well.

This patch also scales the font size inside tooltips. When zooming in,
the font size grows similarly to the font inside the stack frames. There
are some arbitrary upper and lower bounds on the font size, trying to
keep it large enough to be readable on low zoom levels, and small enough
to fit on a screen on high zoom levels.