Add option for hidden stack frame alignment
In 0x, hidden stack frames still take up space. That's visible as empty
space between different stacks: one of the stacks is a child of a hidden
stack frame. This is the easiest thing to do d3 wise and has the nice
benefit that stack frames only move vertically when you toggle frame
visibility.

This PR adds a `collapseHiddenNodeWidths: true` option to change that
behaviour. Hidden stack frames don't take up horizontal space when this
option is set. That means that child stacks of hidden stack frames are
bunched up next to each other without empty space in between. The
benefit of this approach is that, combined with `heatBars: true`, the
heat bars will be one contiguous block, and the size of this block is
directly proportional to the time that frame was at the top of the stack.
Without this option, heat bars are seen through all the gaps where
hidden stack frames are taking up space, and only the colour is
proportional to the heat.

It's quite a mouthful for a fairly small code change...